% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auth.R
\name{app_set_up}
\alias{app_set_up}
\title{App Set Up}
\usage{
app_set_up(app_name = "google")
}
\arguments{
\item{app_name}{app would you like to authorize? Supported apps are 'google' 'calendly' and 'github'}
}
\description{
This is a function that sets up the app. It's generally called by another function
}
