% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_data.R
\name{bad_quiz_path}
\alias{bad_quiz_path}
\title{Path to bad example quiz}
\usage{
bad_quiz_path()
}
\value{
The file path to an example bad quiz included in the package that will fail the quiz checks.
}
\description{
Path to bad example quiz
}
\examples{

quiz_path <- bad_quiz_path()
}
