% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bookdown_to_leanpub.R
\name{bookdown_path}
\alias{bookdown_path}
\title{Find main Bookdown directory}
\usage{
bookdown_path(path = ".")
}
\arguments{
\item{path}{Where to look for the file. By default looks in current directory.}
}
\value{
Returns the directory where the _bookdown.yml is contained.
}
\description{
Find main Bookdown directory
}
