% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bookdown_to_leanpub.R
\name{bookdown_rmd_files}
\alias{bookdown_rmd_files}
\title{Get file paths all Rmds in the bookdown directory}
\usage{
bookdown_rmd_files(path = ".")
}
\arguments{
\item{path}{Where to look for the _bookdown.yml file. Passes toget_bookdown_spec() function. By default looks in current directory}
}
\value{
The file paths to Rmds listed in the _bookdown.yml file.
}
\description{
Get file paths all Rmds in the bookdown directory
}
