% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bookdown_to_leanpub.R
\name{bookdown_to_book_txt}
\alias{bookdown_to_book_txt}
\title{Convert Bookdown to Leanpub}
\usage{
bookdown_to_book_txt(path = ".", output_dir = "manuscript", verbose = TRUE)
}
\arguments{
\item{path}{path to the bookdown book, must have a `_bookdown.yml` file}

\item{output_dir}{output directory to put files.  It should likely be
relative to path}

\item{verbose}{print diagnostic messages}
}
\value{
A list of output files in order, the book text file name, and diagnostics
}
\description{
Convert Bookdown to Leanpub
}
