% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bookdown_to_leanpub.R
\name{bookdown_to_leanpub}
\alias{bookdown_to_leanpub}
\title{Convert Bookdown to Leanpub}
\usage{
bookdown_to_leanpub(
  path = ".",
  render = TRUE,
  output_dir = "manuscript",
  remove_resources_start = FALSE,
  verbose = TRUE
)
}
\arguments{
\item{path}{path to the bookdown book, must have a `_bookdown.yml` file}

\item{render}{if `TRUE`, then [bookdown::render_book()] will be run on each Rmd.}

\item{output_dir}{output directory to put files.  It should likely be
relative to path}

\item{remove_resources_start}{remove the word `resources/` at the front
of any image path.}

\item{verbose}{print diagnostic messages}
}
\value{
A list of output files and diagnostics
}
\description{
Convert Bookdown to Leanpub
}
