% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/token-handlers.R
\name{cache_secrets_folder}
\alias{cache_secrets_folder}
\title{See where your cached secrets are being stored}
\usage{
cache_secrets_folder()
}
\value{
an file path that shows where your cached secrets are stored
}
\description{
This is a function to retrieve the file path of where your cached secrets are stored
}
\examples{
\dontrun{

# You can see where your cached secrets are being stored by running:
cached_secrets_folder()
}
}
