% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auth.R
\name{check_auth}
\alias{check_auth}
\title{Check authentication for rgoogleslides}
\usage{
check_auth(...)
}
\arguments{
\item{...}{Arguments passed to _auth}
}
\description{
Check authentication for rgoogleslides
}
