% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auth.R
\name{check_didactr_auth}
\alias{check_didactr_auth}
\title{Check authentication for rgoogleslides}
\usage{
check_didactr_auth(...)
}
\arguments{
\item{...}{Arguments passed to didactr_auth}
}
\description{
Check authentication for rgoogleslides
}
