% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quiz.R
\name{check_question}
\alias{check_question}
\title{Check Quiz Question Set Up}
\usage{
check_question(question_df, quiz_name = NULL, verbose = TRUE)
}
\arguments{
\item{question_df}{Which is an individual question's data.frame after being parse from}

\item{quiz_name}{The name of the quiz the question is from}

\item{verbose}{Whether progress messages should be given}
}
\value{
A list of messages/warnings regarding each check for the given question.
}
\description{
Check quiz question set up to see if it is compliant with Leanpub and Coursera needs.
Based off of [Markua guide](https://leanpub.com/markua/read#leanpub-auto-quizzes-and-exercises)
}
