% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quiz.R
\name{check_quiz}
\alias{check_quiz}
\title{Check Quizzes}
\usage{
check_quiz(quiz_path, verbose = TRUE)
}
\arguments{
\item{quiz_path}{a full path to a quiz markdown file}

\item{verbose}{print diagnostic messages}
}
\value{
A list of logical indicators
}
\description{
Check Quizzes
}
