% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quiz.R
\name{check_quiz_attributes}
\alias{check_quiz_attributes}
\title{Check Quiz Attributes}
\usage{
check_quiz_attributes(quiz_specs, quiz_name = NULL, verbose = TRUE)
}
\arguments{
\item{quiz_specs}{The output from [leanbuild::parse_quiz].}

\item{quiz_name}{A character string indicating the name of the quiz being checked.}

\item{verbose}{Would you like progress messages? TRUE/FALSE}
}
\value{
A logical
}
\description{
Check Quiz Attributes
}
\examples{

quiz <- c(
  "{quiz, id: quiz_00_filename, choose-answers: 4}",
  "### Lesson Name quiz",
  "{choose-answers: 4, attempts: 25}",
  "? What do you think?",
  "C) The answer to this one",
  "o) Not the answer",
  "o) Not the answer either",
  "C) Another correct answer",
  "m) Mandatory different answer",
  "{/quiz}"
)
quiz_specs <- parse_quiz(quiz)
check_quiz_attributes(quiz_specs)
check_quiz_question_attributes(quiz_specs)

quiz <- c(
  "{quiz, id: quiz_00_filename, choose-answers: 4}",
  "### Lesson Name quiz",
  "{choose-answers: 4, attempts: 25}",
  "",
  "? What do you think?",
  "! The answer to this one",
  "{/quiz}"
)
quiz_specs <- parse_quiz(quiz)
check_quiz_attributes(quiz_specs)
check_quiz_question_attributes(quiz_specs)


}
