% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quiz.R
\name{check_quizzes}
\alias{check_quizzes}
\title{Check all quizzes in a directory}
\usage{
check_quizzes(quiz_dir = "quizzes", write_report = TRUE, verbose = TRUE)
}
\arguments{
\item{quiz_dir}{A path to a directory full of quizzes that should all be checked with [ottr_aide::check_all_quizzes].}

\item{write_report}{TRUE/FALSE save warning report to a CSV file?}

\item{verbose}{print diagnostic messages}
}
\value{
A list checks performed on each quiz
}
\description{
Check the formatting of all quizzes in a given directory.
}
\examples{


## Make a temporary quiz directory
quiz_dir <- dirname(good_quiz_path())

## Now check the quizzes in that directory
all_quiz_results <- check_quizzes(quiz_dir = quiz_dir)
}
