% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_structure.R
\name{check_structure}
\alias{check_structure}
\title{Check for expected file structure and create empty folders if necessary}
\usage{
check_structure(
  course_dir = ".",
  book_txt = NULL,
  verbose = TRUE,
  warn_book_exists = TRUE,
  create_directories = TRUE
)
}
\arguments{
\item{course_dir}{directory with course materials}

\item{book_txt}{A text file with names of markdown
files and course names}

\item{verbose}{print diagnostic messages}

\item{warn_book_exists}{Should a warning be sent if \code{Book.txt}
already exists.}

\item{create_directories}{if the directories do not exist,
create them.}
}
\value{
A list of the paths necessary for the course
}
\description{
Check for expected file structure and create empty folders if necessary
}
\examples{
course_dir = tempfile()
dir.create(course_dir)
check_structure(course_dir)
}
