% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_data.R
\name{clean_up}
\alias{clean_up}
\title{Clean up OTTR_Template files used for testing}
\usage{
clean_up()
}
\value{
Looks for dangling zips and directories downloaded for testing and removes them
}
\description{
Clean up OTTR_Template files used for testing
}
