% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quiz_formatting.R
\name{convert_leanpub_quizzes}
\alias{convert_leanpub_quizzes}
\title{Convert directory of Coursera yaml quizzes to Leanpub md quizzes}
\usage{
convert_leanpub_quizzes(
  input_quiz_dir = "coursera_quizzes",
  output_quiz_dir = "quizzes",
  verbose = TRUE
)
}
\arguments{
\item{input_quiz_dir}{Directory containing Coursera .yml/.yaml quiz files.
Default is \code{"coursera_quizzes"}.}

\item{output_quiz_dir}{Directory where Leanpub .md files will be saved.
Created if it does not exist. Default is \code{"quizzes"}.}

\item{verbose}{Whether to print progress messages.}
}
\value{
Invisibly, the vector of output file paths.
}
\description{
Converts all .yml/.yaml quiz files in a directory to Leanpub .md format using
[convert_quiz_to_leanpub].
}
\examples{
\dontrun{
tdir <- tempfile()
dir.create(tdir, showWarnings = FALSE, recursive = TRUE)
convert_quiz(good_quiz_path(), tdir)
convert_leanpub_quizzes(input_quiz_dir = tdir, output_quiz_dir = "quizzes")
}
}
