% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coursera.R
\name{convert_quiz}
\alias{convert_quiz}
\title{Convert Leanpub md quiz to Coursera yaml quiz}
\usage{
convert_quiz(quiz_path, output_quiz_dir = dirname(quiz_path), verbose = TRUE)
}
\arguments{
\item{quiz_path}{A path to a quiz .md file to be converted.}

\item{output_quiz_dir}{An existing folder where you would like the new version of the quiz to be saved.
Default is the directory of the quiz_path provided}

\item{verbose}{Would you like the progress messages?}
}
\value{
A Coursera-ready quiz file saved to the output directory specified as a yaml.
}
\description{
Convert a Leanpub-formatted md quiz file to a Coursera-formatted yaml quiz file in preparation for uploading to Coursera.
}
\examples{

quiz_path <- good_quiz_path()

# Provide path to quiz to convert
convert_quiz(quiz_path)


}
