% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quiz_formatting.R
\name{convert_quiz_to_leanpub}
\alias{convert_quiz_to_leanpub}
\title{Convert Coursera yaml quiz to Leanpub md quiz}
\usage{
convert_quiz_to_leanpub(
  quiz_path,
  output_quiz_dir = dirname(quiz_path),
  verbose = TRUE
)
}
\arguments{
\item{quiz_path}{Path to a Coursera .yml or .yaml quiz file.}

\item{output_quiz_dir}{Directory where the Leanpub .md file will be saved.
Default is the directory of \code{quiz_path}.}

\item{verbose}{Whether to print progress messages.}
}
\value{
Path to the written Leanpub quiz file (invisibly).
}
\description{
Convert a Coursera-formatted yaml quiz file (as produced by [convert_quiz])
to a Leanpub-formatted md quiz file.
}
\examples{
\dontrun{
# Convert a Coursera quiz to Leanpub
yml_path <- system.file("extdata", "quizzes", "quiz_coursera_sample.yml", package = "ottrpal")
if (nzchar(yml_path)) convert_quiz_to_leanpub(yml_path)
}
}
