% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/replace_html.R
\name{convert_utube_link}
\alias{convert_utube_link}
\title{Convert youtube link}
\usage{
convert_utube_link(utube_link)
}
\arguments{
\item{utube_link}{a link to a youtube video that may or may not be "www.youtube.com/embed" or "www.youtube.com/watch?v="
format.}
}
\value{
Returns a youtube link in the "watch" format so it will render properly in Leanpub or Coursera-friendly files
}
\description{
Convert youtube link
}
