% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bookdown_to_leanpub.R
\name{copy_directory_contents}
\alias{copy_directory_contents}
\title{Copying directory contents}
\usage{
copy_directory_contents(from, to)
}
\arguments{
\item{from}{Where the files to copy are located.}

\item{to}{Where the files to copy are to be copied should go to.}
}
\value{
The file paths to Rmds listed in the _bookdown.yml file.
}
\description{
Copying directory contents
}
