% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_up.R
\name{course_path}
\alias{course_path}
\title{Find main course git directory}
\usage{
course_path(path = ".")
}
\arguments{
\item{path}{Where to look for the file. By default looks in current directory.}
}
\value{
Returns the directory where the .git folder is contained.
}
\description{
Find main course git directory
}
