% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coursera.R
\name{convert_quiz}
\alias{convert_quiz}
\title{Convert Leanpub md quiz to Coursera yaml quiz}
\usage{
convert_quiz(quiz_path, output_quiz_dir, verbose = TRUE)
}
\arguments{
\item{quiz_path}{A path to a quiz .md file.}

\item{output_quiz_dir}{An existing folder where you would like the new version of the quiz to be saved.}

\item{verbose}{Would you like the progress messages?}
}
\value{
A Coursera-ready quiz file saved to the output directory specified as a yaml.
}
\description{
Convert Leanpub md quiz to Coursera yaml quiz
}
