% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{download_ottr_template}
\alias{download_ottr_template}
\title{Download files from main OTTR_Template to test}
\usage{
download_ottr_template(dir = "inst/extdata", type = "rmd")
}
\arguments{
\item{dir}{What relative file path should the files be downloaded}

\item{type}{Which OTTR repo are we downloading? Options are "rmd", "quarto", "rmd_website", "quarto_website"}
}
\value{
This downloads the main branch repo files from the respective repo for testing purposes
}
\description{
Download files from main OTTR_Template to test
}
