% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_data.R
\name{encrypt_creds_user_path}
\alias{encrypt_creds_user_path}
\title{Get file path to an default credentials RDS}
\usage{
encrypt_creds_user_path()
}
\description{
Get file path to an default credentials RDS
}
