% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/github_handling.R
\name{find_issue}
\alias{find_issue}
\title{Find an issue on GitHub with a particular title}
\usage{
find_issue(text, repo_name, token = NULL)
}
\arguments{
\item{text}{What text to be searched for in the GitHub issues. Can be regex.}

\item{repo_name}{the name of the repository, e.g. jhudsl/OTTR_Template}

\item{token}{A personal access token from GitHub. Only necessary if the
repository being checked is a private repository.}
}
\value{
A TRUE/FALSE whether or not the issue with this text on this repository exists.
}
\description{
Given text and repository name, find if an issue exists.
}
\examples{
\dontrun{

authorize("github")

find_issue(text = "TEST", repo_name = "jhudsl/ottrpal")

}
}
