% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/footnotes.R
\name{convert_footnotes}
\alias{convert_footnotes}
\title{Convert Bookdown footnotes to Leanpub-formatted footnotes}
\usage{
convert_footnotes(content)
}
\arguments{
\item{content}{a character vector containing the lines of content from a file read in with readLines()}
}
\value{
a character vector containing the content given but with Leanpub formatted foonotes
}
\description{
Convert Bookdown footnotes to Leanpub-formatted footnotes
}
