% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bookdown_to_leanpub.R
\name{get_bookdown_spec}
\alias{get_bookdown_spec}
\title{Load in Bookdown specifications from _bookdown.yml}
\usage{
get_bookdown_spec(path = ".")
}
\arguments{
\item{path}{Where to look for the _bookdown.yml file. Passes to the bookdown_file() function. By default looks in current directory}
}
\value{
The yaml contents using yaml::yaml.load_file()
}
\description{
Load in Bookdown specifications from _bookdown.yml
}
