% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/github_handling.R
\name{get_git_auth}
\alias{get_git_auth}
\title{Handle GitHub PAT authorization}
\usage{
get_git_auth(
  git_pat = NULL,
  git_username = "PersonalAccessToken",
  quiet = FALSE
)
}
\arguments{
\item{git_pat}{If private repositories are to be retrieved, a github personal
access token needs to be supplied. If none is supplied, then this will attempt to
grab from a git pat set in the environment with usethis::create_github_token().}

\item{git_username}{Optional, can include username for credentials.}

\item{quiet}{Use TRUE if you don't want the warning about no GitHub credentials.}
}
\value{
Authorization argument to supply to curl OR a blank string if no
authorization is found or supplied.
}
\description{
Handle things whether or not a GitHub PAT is supplied.
}
