% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/notes_to_fig_alt.R
\name{get_gs_pptx}
\alias{get_gs_pptx}
\title{Download Google Slides pptx file}
\usage{
get_gs_pptx(id)
}
\arguments{
\item{id}{Identifier of Google slides presentation, passed to
\code{\link{get_slide_id}}}
}
\value{
Downloaded file (in temporary directory)
}
\description{
Download Google Slides pptx file
}
\note{
This downloads presentations if they are public and also try to make
sure it does not fail on large files
}
