% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/notes_to_fig_alt.R
\name{get_gs_slide_pptx}
\alias{get_gs_slide_pptx}
\title{Download a slide from Google Slides as PPTX}
\usage{
get_gs_slide_pptx(url)
}
\arguments{
\item{id}{Identifier of Google slides presentation, passed to
\code{\link{get_presentation_id}}}
}
\value{
Downloaded file (in temporary directory)
}
\description{
Download a slide from Google Slides as PPTX
}
\note{
This downloads presentations if they are public and also try to make
sure it does not fail on large files
}
