% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/notes_to_fig_alt.R
\name{get_object_id_notes}
\alias{get_object_id_notes}
\title{Retrieve Speaker Notes and their corresponding Object (Slide) IDs from a Google Slides presentation}
\usage{
get_object_id_notes(slide_url)
}
\arguments{
\item{slide_url}{URL whose 'General access' is set to 'Anyone with the link'}
}
\value{
Data frame of Object IDs and Speaker notes.
}
\description{
Google Slides API calls a presentation slide ID as an 'object ID'.
}
\examples{
\dontrun{
 get_object_id_notes("https://docs.google.com/presentation/d/
                      1H5aF_ROKVxE-HFHhoOy9vU2Y-y2M_PiV0q-JBL17Gss/edit?usp=sharing")
}
}
