% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gs_png.R
\name{get_presentation_id}
\alias{get_presentation_id}
\title{Get Presentation ID from URL}
\usage{
get_presentation_id(x)
}
\arguments{
\item{x}{URL of slide}
}
\value{
A character vector
}
\description{
Get Presentation ID from URL
}
\examples{
x <- paste0(
  "https://docs.google.com/presentation/d/",
  "1Tg-GTGnUPduOtZKYuMoelqUNZnUp3vvg_7TtpUPL7e8",
  "/edit#slide=id.g154aa4fae2_0_58"
)
get_presentation_id(x)
}
