% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/notes_to_fig_alt.R
\name{get_slide_urls}
\alias{get_slide_urls}
\title{Collect all the google slide URLs from the docs folder}
\usage{
get_slide_urls(
  path = ".",
  pattern = "1YmwKdIy9BeQ3EShgZhvtb3MgR8P6iDX4DfFD65W_gdQ",
  docs = "docs",
  file_suffix = "html$",
  output_dir = "resources",
  resources_dir = "resources",
  exclude_file = ""
)
}
\arguments{
\item{path}{path to the bookdown or quarto course repository, must have a
`.github` folder which will be used to establish the top of the repo.}

\item{pattern}{What is the pattern to be used to identify the strings in the html files?}

\item{docs}{What is the file path to the docs folder?}

\item{file_suffix}{What types of files should we be looking in?}

\item{output_dir}{A relative file path to the folder (existing or not) that the
output check file should be saved to. Default is "check_reports"}

\item{resources_dir}{A relative file path to the folder (existing or not) that the
ignore_urls.txt file and exclude_files.txt will be found. Default is "resources".
If no ignore_urls.txt file and exclude_files.txt files are found, we will download one.}
}
\value{
a file and data.frame of all the google slide URLs identified in the given html docs files
}
\description{
Collect all the google slide URLs from the docs folder
}
\examples{
\dontrun{

rmd_dir <- setup_ottr_template(dir = ".", type = "rmd", render = FALSE)
slide_deck <- "1YmwKdIy9BeQ3EShgZhvtb3MgR8P6iDX4DfFD65W_gdQ"
slide_urls <- get_slide_urls(rmd_dir, pattern = slide_deck)
}
}
