% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_up.R
\name{get_yaml_spec}
\alias{get_yaml_spec}
\title{Load in yaml specifications from _bookdown.yml or _quarto.yml}
\usage{
get_yaml_spec(path = ".")
}
\arguments{
\item{path}{Where to look for the yaml spec file. By default looks in current directory.}
}
\value{
The yaml contents using yaml::yaml.load_file()
}
\description{
Load in yaml specifications from _bookdown.yml or _quarto.yml
}
