% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_data.R
\name{good_quiz_path}
\alias{good_quiz_path}
\title{Path to good example quiz}
\usage{
good_quiz_path()
}
\value{
The file path to an example good quiz included in the package that should pass the quiz checks.
}
\description{
Path to good example quiz
}
