% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/google_slides.R
\name{gs_id_from_slide}
\alias{gs_id_from_slide}
\title{Google Slides Helper Functions}
\usage{
gs_id_from_slide(file)
}
\arguments{
\item{file}{markdown file for manuscript}
}
\value{
A scalar character vector
}
\description{
Google Slides Helper Functions
}
