% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/leanpub_checks.R
\name{leanpub_check}
\alias{leanpub_check}
\title{Check Leanpub Course or Book}
\usage{
leanpub_check(path = ".", verbose = TRUE)
}
\arguments{
\item{path}{path to the Leanpub book/course}

\item{verbose}{print diagnostic messages}
}
\value{
A list of output files and diagnostics
}
\description{
Check Leanpub Course or Book
}
