% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_bookdown.R
\name{lp_convert_bookdown}
\alias{lp_convert_bookdown}
\title{Convert Bookdown Book to Leanpub}
\usage{
lp_convert_bookdown(path = ".", bookdown_dir = NULL)
}
\arguments{
\item{path}{path to the directory with bookdown}

\item{bookdown_dir}{Directory for output of bookdown run.  If
\code{NULL}, then  `output_dir` from `_bookdown.yml` will be extracted.
Otherwise, defaults to `docs`}
}
\value{
A character vector of output files
}
\description{
Convert Bookdown Book to Leanpub
}
