% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{markdown_quiz_path}
\alias{markdown_quiz_path}
\title{Get file path to an example quiz}
\usage{
markdown_quiz_path()
}
\value{
A file path to a markua markdown quiz example you can use for testing
}
\description{
Get file path to an example quiz
}
\examples{
\dontrun{

# Find quiz path

quiz_path <- markdown_quiz_path()
}

}
