% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{ottr_check}
\alias{ottr_check}
\title{Ottrpal Checks}
\usage{
ottr_check(
  check_type,
  path = ".",
  output_dir = "check_reports",
  resources_dir = "resources",
  ...
)
}
\arguments{
\item{check_type}{What check would you like to run "spelling", "urls" or
"quiz_format".}

\item{path}{path to the bookdown or quarto course repository, must have a
`.github` folder which will be used to establish the top of the repo.}

\item{output_dir}{A relative file path to the folder (existing or not) that the
output check file should be saved to. Default is "check_reports"}

\item{resources_dir}{A relative file path to the folder (existing or not) that the
dictionary.txt file and exclude_files.txt will be found. Default is "resources".
If no dictionary.txt file and exclude_files.txt files are found, we will download one.}

\item{...}{All additional arguments passed to the respective check}
}
\value{
The result of the check being called
}
\description{
Ottrpal Checks
}
