% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ottrfy.R
\name{ottrfy}
\alias{ottrfy}
\title{OTTRfy your repository}
\usage{
ottrfy(path = ".", type = "rmd", git_commit = TRUE, overwrite = FALSE)
}
\arguments{
\item{path}{What's the file path we are making an OTTR. Needs to be a directory}

\item{type}{Can be "rmd", "quarto" "rmd_web", or "quarto_web" depending on what
kind of OTTR site you'd like to make}
}
\value{
Information regarding a Github account
}
\description{
This script downloads all the files and sets up the folders you need to
OTTR-fy a repository that has markdown or R Markdown files
}
\examples{
\dontrun{

ottrfy(type = "rmd")

ottrfy(type = "quarto")

ottrfy(type = "rmd_web")

ottrfy(type = "quarto_web")

}

}
