% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_up.R
\name{output_destination}
\alias{output_destination}
\title{Declare file path to docs/ folder}
\usage{
output_destination(path = ".")
}
\arguments{
\item{path}{Where to look for the _bookdown.yml or _quarto.yml file. Passes to get_yaml_spec() function. By default looks in current directory}
}
\value{
The file paths to rmds or qmds listed in the _bookdown.yml or _quarto.ymlfile.
}
\description{
Declare file path to docs/ folder
}
