% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quiz_formatting.R
\name{parse_q_tag}
\alias{parse_q_tag}
\title{Parse apart a tag}
\usage{
parse_q_tag(tag)
}
\arguments{
\item{tag}{A single tag to extract from}
}
\value{
A named vector indicating the field and entry associated with it.
}
\description{
Parse apart a tag
}
\examples{

tag <- "{quiz, id: quiz_name_here, attempts: 10}"
parse_q_tag(tag)
}
