% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quiz.R
\name{parse_quiz}
\alias{parse_quiz}
\alias{extract_quiz}
\alias{find_quiz_indices}
\title{Parse Quiz}
\usage{
parse_quiz(x)

extract_quiz(x)

find_quiz_indices(x)
}
\arguments{
\item{x}{A single filename or a vector of the contents of the markdown
file}
}
\value{
A list of elements, including a `data.frame` and metadata
for questions
}
\description{
Parse Quiz
}
\examples{


}
