% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quiz.R
\name{parse_quiz_df}
\alias{parse_quiz_df}
\title{Parse quiz into a data.frame}
\usage{
parse_quiz_df(quiz_lines, remove_tags = FALSE)
}
\arguments{
\item{quiz_lines}{A character vector of the contents of the markdown
file obtained from readLines()}

\item{remove_tags}{TRUE/FALSE remove tags and empty lines?}
}
\value{
A data frame containing a type column which indicates what type of line each is.
}
\description{
Parse quiz into a data.frame
}
\examples{

# Use readLines() to read in a quiz
quiz_path <- good_quiz_path()
quiz_lines <- readLines(quiz_path)

# Can use this to parse the quiz into a data.frame
quiz_df <- parse_quiz_df(quiz_lines)
}
