% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{\%>\%}
\alias{\%>\%}
\title{Pipe operator}
\usage{
lhs \%>\% rhs
}
\arguments{
\item{lhs}{A value or the dplyr placeholder.}

\item{rhs}{A function call using the dplyr semantics.}
}
\value{
The result of calling `rhs(lhs)`.
}
\description{
Pipe operator
}
\keyword{internal}
