% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_up.R
\name{qrmd_files}
\alias{qrmd_files}
\title{Get file paths to all qmds or rmds in the course website directory}
\usage{
qrmd_files(path = ".")
}
\arguments{
\item{path}{Where to look for the _bookdown.yml or _quarto.yml file. Passes to get_yaml_spec() function. By default looks in current directory}
}
\value{
The file paths to rmds or wmds listed in the _bookdown.yml or _quarto.yml file.
}
\description{
Get file paths to all qmds or rmds in the course website directory
}
