% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quiz.R
\docType{data}
\name{quiz}
\alias{quiz}
\title{Parse Quiz and Other Checking Functions}
\format{
An object of class \code{character} of length 1.
}
\usage{
quiz
}
\arguments{
\item{quiz}{A single filename or a vector of the contents of the markdown
file}
}
\value{
A list of elements, including a `data.frame` and metadata
for questions
}
\description{
Parse Quiz and Other Checking Functions
}
\examples{

quiz <- c(
  "{quiz, id: quiz_00_filename}",
  "### Lesson Name quiz",
  "{choose-answers: 4}",
  "? What do you think?",
  "",
  "C) The answer to this one",
  "o) Not the answer",
  "o) Not the answer either",
  "C) Another correct answer",
  "m) Mandatory different answer",
  "",
  "{/quiz}"
)
out <- parse_quiz(quiz)
check_quiz_attributes(out)
}
\keyword{datasets}
