% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remove_yaml.R
\name{remove_yaml_header}
\alias{remove_yaml_header}
\title{Remove YAML header}
\usage{
remove_yaml_header(file)
}
\arguments{
\item{file}{file name of the markdown file}
}
\value{
A character vector of the text without the YAML header
}
\description{
Remove YAML header
}
\examples{
file <- system.file("extdata/00_template.Rmd", package = "ottrpal")
out <- remove_yaml_header(file)
head(out)
}
