% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/replace_html.R
\name{replace_html}
\alias{replace_html}
\alias{replace_single_html}
\title{Replace HTML and other Tags in Leanpub Markdown}
\usage{
replace_html(path = "manuscript", fullbleed = FALSE, verbose = TRUE)

replace_single_html(file, fullbleed = FALSE)
}
\arguments{
\item{path}{path to the markdown files that need replacement.}

\item{fullbleed}{should the image have the attribute `fullbleed: true`?}

\item{verbose}{print diagnostic messages}

\item{file}{individual markdown file}
}
\value{
A list of output files and diagnostics
}
\description{
Replace HTML and other Tags in Leanpub Markdown
}
