% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auth.R
\name{rgoogle_auth}
\alias{rgoogle_auth}
\title{Authorize Application}
\usage{
rgoogle_auth(
  token_file = NULL,
  cache = FALSE,
  language = FALSE,
  use_oob = FALSE,
  token = NULL
)
}
\arguments{
\item{token_file}{If the \code{token} has been saved, use this file
to load the credentials.}

\item{cache}{A logical value or a string. \code{TRUE} means to cache
using the default cache file \code{.httr-oauth}}

\item{language}{Should the language API be authorized
using your token (experimental, and most likely doesn't work)}

\item{use_oob}{use a local webserver for the OAuth dance}

\item{token}{you can pass the token directly instead of the file
if you want}
}
\value{
The auth token, a Token class.
}
\description{
Authorize Application
}
\examples{
\dontrun{
rgoogle_auth()
}
}
