% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_up.R
\name{set_up_leanpub}
\alias{set_up_leanpub}
\title{Set up Manuscript folder for Leanpub publishing}
\usage{
set_up_leanpub(
  path = ".",
  clean_up = FALSE,
  render = NULL,
  output_dir = "manuscript",
  make_book_txt = FALSE,
  quiz_dir = "quizzes",
  run_quiz_checks = FALSE,
  remove_resources_start = FALSE,
  verbose = TRUE,
  footer_text = NULL
)
}
\arguments{
\item{path}{path to the top of course repository (looks for .git folder)}

\item{clean_up}{TRUE/FALSE the old output directory should be deleted and
everything created fresh.}

\item{render}{If NULL will not be run. If "quarto" or "bookdown" then the respective render type will be run}

\item{output_dir}{output directory to put files.  It should likely be
relative to path}

\item{make_book_txt}{Should [ottrpal::course_to_book_txt()] be run
to create a `Book.txt` in the output directory?}

\item{quiz_dir}{directory that contains the quiz .md files that should be
checked and incorporated into the Book.txt file. If you don't have quizzes,
set this to NULL}

\item{run_quiz_checks}{TRUE/FALSE run quiz checks}

\item{remove_resources_start}{remove the word `resources/` at the front
of any image path.}

\item{verbose}{print diagnostic messages}

\item{footer_text}{Optionally can add a bit of text that will be added to the
end of each file before the references section.}
}
\value{
A list of output files and diagnostics
}
\description{
Set up Manuscript folder for Leanpub publishing
}
