% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_data.R
\name{setup_ottr_template}
\alias{setup_ottr_template}
\title{Download and render files from main OTTR_Template to test}
\usage{
setup_ottr_template(dir = "inst/extdata", type)
}
\arguments{
\item{dir}{What relative file path should the files be downloaded}

\item{type}{Which OTTR repo are we downloading? Options are "rmd", "quarto", "rmd_website", "quarto_website"}
}
\value{
This downloads the main branch repo files from the respective repo for testing purposes
}
\description{
Download and render files from main OTTR_Template to test
}
