% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simple_references.R
\name{simple_references}
\alias{simple_references}
\title{Run Simple References}
\usage{
simple_references(x, bib_files, add_reference_header = FALSE)
}
\arguments{
\item{x}{A filename of a markdown or Rmarkdown file, or the full output
from `readLines` on that file}

\item{bib_files}{bibliography files to use for pandoc}

\item{add_reference_header}{Should the `## References` header be added at the
end of the output?}
}
\value{
A character vector of the file, with references subbed in
}
\description{
Run Simple References
}
