% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/url_check.R
\name{test_url}
\alias{test_url}
\title{Test a URL}
\usage{
test_url(url, ignore_urls = "")
}
\arguments{
\item{url}{A single URL that will be checked whether it is real.}

\item{ignore_url}{A vector of URLs which to ignore.}
}
\value{
a logical TRUE/FALSE for whether the URL is legitimate.
}
\description{
Test a URL
}
\examples{
/dontrun {

# This should print out "failed"
test_url("https://notawebsiteaaaaaaa.com")


# This should print out "success"
test_url("https://github.com")
}

}
