% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ottrfy.R
\name{update_chapters}
\alias{update_chapters}
\title{Update chapters}
\usage{
update_chapters(path = ".", ignore_files = NULL)
}
\arguments{
\item{path}{What's the file path to where we are updating chapters?}

\item{ignore_files}{Are there qmd or rmd files that are not chapters and should be ignored?}
}
\value{
an updated _bookdown.yml or _quarto.yml with all chapters.
}
\description{
This updates the chapter list in the _bookdown.yml or _quarto.yml
Default behavior is to sort the chapters but have index come first.
}
\examples{
\dontrun{

update_chapters()

}

}
