% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bookdown_to_leanpub.R
\name{bookdown_destination}
\alias{bookdown_destination}
\title{Declare file path to docs/ folder}
\usage{
bookdown_destination(path = ".")
}
\arguments{
\item{path}{Where to look for the _bookdown.yml file. Passes toget_bookdown_spec() function. By default looks in current directory}
}
\value{
The file paths to Rmds listed in the _bookdown.yml file.
}
\description{
Declare file path to docs/ folder
}
