% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_up.R
\name{bookdown_file}
\alias{bookdown_file}
\title{Find file path to _bookdown.yml}
\usage{
bookdown_file(path = ".")
}
\arguments{
\item{path}{Where to look for the _bookdown.yml file.
Passes to the [bookdown_file()] function. By default looks in current directory}
}
\value{
The file path to `_bookdown.yml`
}
\description{
Find file path to _bookdown.yml
}
