% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bookdown_to_leanpub.R
\name{bookdown_to_book_txt}
\alias{bookdown_to_book_txt}
\title{Create Book.txt file from files existing in quiz directory}
\usage{
bookdown_to_book_txt(
  path = ".",
  md_files = NULL,
  output_dir = "manuscript",
  quiz_dir = "quizzes",
  verbose = TRUE
)
}
\arguments{
\item{path}{path to the bookdown book, must have a `_bookdown.yml` file}

\item{md_files}{vector of file path of the md's to be included}

\item{output_dir}{output directory to put files.  It should likely be
relative to path}

\item{quiz_dir}{Where are the quizzes stored? Default looks for folder called "quizzes".}

\item{verbose}{print diagnostic messages}
}
\value{
A list of quiz and chapter files in order in a file called Book.txt -- How Leanpub wants it.
}
\description{
Create Book.txt file from files existing in quiz directory
}
