% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quiz.R
\name{check_all_questions}
\alias{check_all_questions}
\title{Check All Quiz Questions}
\usage{
check_all_questions(quiz_specs, quiz_name = NULL, verbose = TRUE)
}
\arguments{
\item{quiz_specs}{quiz_specs which is output from [leanbuild::parse_quiz]}

\item{quiz_name}{The name of the quiz being checked.}

\item{verbose}{Whether progress messages should be given.}
}
\value{
A list of the output from [leanbuild::check_question] with messages/warnings regarding each question and each check.
}
\description{
Check All Quiz Questions
}
\examples{
quiz <- c(
  "{quiz, id: quiz_00_filename, choose-answers: 4}",
  "### Lesson Name quiz",
  "{choose-answers: 4, attempts: 25}",
  "? What do you think?",
  "C) The answer to this one",
  "o) Not the answer",
  "o) Not the answer either",
  "C) Another correct answer",
  "m) Mandatory different answer",
  "{/quiz}"
)
quiz_specs <- parse_quiz(quiz)
quiz_checks <- check_all_questions(quiz_specs)

}
