% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quiz.R
\name{check_quiz_attributes}
\alias{check_quiz_attributes}
\alias{check_quiz_answers}
\title{Check Quiz Information}
\usage{
check_quiz_attributes(x)

check_quiz_answers(x, verbose = TRUE)
}
\arguments{
\item{x}{The output from [leanbuild::parse_quiz]}

\item{verbose}{print diagnostic messages}
}
\value{
A logical
}
\description{
Check Quiz Information
}
\examples{

x = c('{quiz, id: quiz_00_filename}',
"### Lesson Name quiz",
"{choose-answers: 4}",
"? What do you think?",
"C) The answer to this one",
"o) Not the answer",
"o) Not the answer either",
"C) Another correct answer",
"m) Mandatory different answer",
"{/quiz}")
out = parse_quiz(x)
check_quiz_attributes(out)
check_quiz_answers(out)
}
