% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quiz.R
\name{check_quiz_attributes}
\alias{check_quiz_attributes}
\title{Check Quiz Attributes}
\usage{
check_quiz_attributes(quiz_specs, quiz_name = NULL, verbose = TRUE)
}
\arguments{
\item{quiz_specs}{The output from [ottr_aide::parse_quiz].}

\item{quiz_name}{A character string indicating the name of the quiz being checked.}

\item{verbose}{Would you like progress messages? TRUE/FALSE}
}
\value{
A logical
}
\description{
Check Quiz Attributes
}
