% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quiz.R
\name{check_quiz_question_attributes}
\alias{check_quiz_question_attributes}
\title{Check a question's attributes}
\usage{
check_quiz_question_attributes(question_df, quiz_name = NULL, verbose = TRUE)
}
\arguments{
\item{question_df}{a data.frame obtained from [leanbuild::parse_quiz_df] and dplyr::group_split(question).}

\item{quiz_name}{inherited from parse}

\item{verbose}{print diagnostic messages}
}
\description{
This is ran automatically by [leanbuild::check_all_questions] for all questions.
It checks that the attributes specified are accepted ones by Leanpub.
}
\examples{

# Use readLines to read in a quiz
quiz_lines <- readLines(good_quiz_path())

# Use
question_df <- parse_quiz_df(quiz_lines) \%>\%
  dplyr::group_split(question)


}
