% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quiz_formatting.R
\name{check_quiz_question_attributes}
\alias{check_quiz_question_attributes}
\title{Check a question's attributes}
\usage{
check_quiz_question_attributes(question_df, quiz_name = NULL, verbose = TRUE)
}
\arguments{
\item{question_df}{a data.frame obtained from [ottrpal::parse_quiz_df] and dplyr::group_split(question).}

\item{quiz_name}{inherited from parse}

\item{verbose}{print diagnostic messages}
}
\value{
Will return a warning for any quiz question attributes used that are not supported.
}
\description{
This is ran automatically by [ottrpal::check_all_questions] for all questions.
It checks that the attributes specified are accepted ones by Leanpub.
}
