% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quiz.R
\name{check_quizzes}
\alias{check_quizzes}
\alias{check_quiz}
\title{Check Quizzes}
\usage{
check_quizzes(path = "manuscript")

check_quiz(path)
}
\arguments{
\item{path}{either a path to the directory of quizzes or a full path to a
quiz markdown file}
}
\value{
A list of logical indicators
}
\description{
Check Quizzes
}
\examples{

x = c('{quiz, id: quiz_00_filename}',
"### Lesson Name quiz",
"{choose-answers: 4}",
"? What do you think?",
"C) The answer to this one",
"o) Not the answer",
"o) Not the answer either",
"C) Another correct answer",
"m) Mandatory different answer",
"{/quiz}")
tdir = tempfile()
dir.create(tdir, showWarnings = FALSE, recursive = TRUE)
tfile = tempfile(pattern = "quiz_", fileext = ".md", tmpdir = tdir)
writeLines(x, tfile)
check_quizzes(path = tdir)

check_quiz(x)
}
