% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quiz.R
\name{check_quizzes}
\alias{check_quizzes}
\title{Check all quizzes in a directory}
\usage{
check_quizzes(path = "quizzes", verbose = TRUE)
}
\arguments{
\item{path}{A path to a directory full of quizzes that should all be checked with [leanbuild::check_all_quizzes].}

\item{verbose}{print diagnostic messages}
}
\value{
A list checks performed on each quiz
}
\description{
Check the formatting of all quizzes in a given directory.
}
\examples{

quiz <- c(
  "{quiz, id: quiz_00_filename}",
  "### Lesson Name quiz",
  "{choose-answers: 4}",
  "? What do you think?",
  "C) The answer to this one",
  "o) Not the answer",
  "o) Not the answer either",
  "C) Another correct answer",
  "m) Mandatory different answer",
  "{/quiz}"
)

## Make a temporary quiz directory
tdir <- tempfile()
dir.create(tdir, showWarnings = FALSE, recursive = TRUE)
tfile <- tempfile(pattern = "quiz_", fileext = ".md", tmpdir = tdir)
writeLines(quiz, tfile)

## Now check the quizzes in that directory
all_quiz_results <- check_quizzes(path = tdir)

}
