% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quiz_formatting.R
\name{check_quizzes}
\alias{check_quizzes}
\title{Check all quizzes in a directory}
\usage{
check_quizzes(
  path = ".",
  quiz_dir = "quizzes",
  write_report = TRUE,
  verbose = TRUE,
  ignore_coursera = TRUE
)
}
\arguments{
\item{path}{path to the top of course repository (looks for .github folder)}

\item{quiz_dir}{A path to a directory full of quizzes that should all be checked with [ottrpal::check_all_quizzes].}

\item{write_report}{TRUE/FALSE save warning report to a CSV file?}

\item{verbose}{print diagnostic messages}

\item{ignore_coursera}{Coursera doesn't like `!` or `:` in the quizzes. Do not convert quizzes to coursera and ignore ! and : in question prompts that would not be allowed in Leanpub quizzes when converted to a Coursera quiz. Default is to ignore Coursera compatibility}
}
\value{
A list checks performed on each quiz
}
\description{
Check the formatting of all quizzes in a given directory.
}
\examples{
\dontrun{


## Make a temporary quiz directory
quiz_dir <- dirname(good_quiz_path())

## Now check the quizzes in that directory
all_quiz_results <- check_quizzes(quiz_dir = quiz_dir)
}
}
