% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{course_path}
\alias{course_path}
\title{Find root of OTTR course provided}
\usage{
course_path(path = ".")
}
\arguments{
\item{path}{Where should we be looking for a OTTR course. By default will look in current working directory.}
}
\value{
Absolute file path to the course pointed to

Returns a absolute file path to where the course is
}
\description{
Find root of OTTR course provided
}
