% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coursera.R
\name{find_end_of_prompt}
\alias{find_end_of_prompt}
\alias{convert_quiz}
\alias{convert_coursera_quizzes}
\alias{render_coursera}
\title{Find the end of the prompt}
\usage{
find_end_of_prompt(start_prompt_index, type_vector)

convert_quiz(quiz_path, output_quiz_dir, verbose = TRUE)

convert_coursera_quizzes(
  input_quiz_dir = "quizzes",
  output_quiz_dir = "coursera_quizzes",
  verbose = TRUE
)

render_coursera(
  output_dir = file.path("docs", "coursera"),
  output_yaml = "_output.yml",
  convert_quizzes = FALSE,
  input_quiz_dir = "quizzes",
  output_quiz_dir = "coursera_quizzes",
  verbose = TRUE
)
}
\arguments{
\item{start_prompt_index}{a single index to start the search at (the beginning of the prompt)}

\item{type_vector}{A vector indicating the type of line -- will look for "answer" to indicate that the prompt has ended.}

\item{quiz_path}{a path to a quiz .md file.}

\item{output_quiz_dir}{A folder (existing or not) where the coursera quizzes should be saved. By default is "coursera_quizzes".}

\item{verbose}{would you like the progress messages?}

\item{input_quiz_dir}{A path to a directory of Leanpub-formatted quiz md files. By default assumes "quizzes" and looks in current directory.}

\item{output_dir}{A folder (existing or not) that the TOC-less Bookdown for Coursera files should be saved. By default is file.path("docs", "coursera")}

\item{output_yaml}{A output.yml file to be provided to bookdown. By default is "_output.yml"}

\item{convert_quizzes}{TRUE/FALSE whether or not to convert quizzes. Default is TRUE}
}
\value{
The index of the end of the prompt

a coursera ready quiz file saved to the output directory specified as a yaml.

a folder of coursera ready quiz files saved to the output directory specified as a yamls.

a folder of coursera ready quiz files saved to the output directory specified as a yamls.
}
\description{
Create a version of Leanpub that does not have a TOC and has quizzes in the Coursera yaml format. Requires Bookdown output files including "assets", "resources", and "libs".
}
