% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coursera.R
\name{render_coursera}
\alias{render_coursera}
\title{Create TOC-less Bookdown for use in Coursera}
\usage{
render_coursera(
  output_dir = file.path("docs", "coursera"),
  convert_quizzes = FALSE,
  input_quiz_dir = "quizzes",
  output_quiz_dir = "coursera_quizzes",
  verbose = TRUE
)
}
\arguments{
\item{output_dir}{a folder (existing or not) that the TOC-less Bookdown for Coursera files should be saved. By default is file.path("docs", "coursera")}

\item{convert_quizzes}{TRUE/FALSE whether or not to convert quizzes. Default is TRUE}

\item{input_quiz_dir}{a path to a directory of leanpub formatted quiz md files. By default assumes "quizzes" and looks in current directory.}

\item{output_quiz_dir}{a folder (existing or not) where the coursera quizzes should be saved. By default is "coursera_quizzes".}

\item{verbose}{would you like the progress messages?}
}
\value{
a folder of coursera ready quiz files saved to the output directory specified as a yamls.
}
\description{
Create a version of Leanpub that does not have a TOC and has quizzes in the Coursera yaml format. Requires Bookdown output files including "assets", "resources", and "libs".
}
