% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example_data.R
\name{example_repo_cleanup}
\alias{example_repo_cleanup}
\title{Clean up example repo files}
\usage{
example_repo_cleanup(dir = ".")
}
\arguments{
\item{dir}{What directory to delete the example files from}
}
\description{
Clean up example repo files
}
\examples{

example_repo_cleanup()

}
