% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example_data.R
\name{example_repo_cleanup}
\alias{example_repo_cleanup}
\title{Clean up example repo files}
\usage{
example_repo_cleanup(files_to_remove, verbose = FALSE)
}
\arguments{
\item{files_to_remove}{List of example files to delete.}

\item{verbose}{TRUE/FALSE would you like progress messages?}
}
\description{
Clean up example repo files
}
\examples{

# Run this to get the files we need
example_files <- ottrpal::example_repo_setup()

# Run this to delete them
example_repo_cleanup(files_to_remove = basename(example_files))
}
