% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example_data.R
\name{example_repo_setup}
\alias{example_repo_setup}
\title{Set up example repo files}
\usage{
example_repo_setup(dest_dir = ".")
}
\arguments{
\item{dest_dir}{The destination directory you would like the example repo files to be placed. By default is current directory.}
}
\description{
Set up example repo files
}
\examples{
# Run this to get the files we need
example_files <- leanbuild::example_repo_setup()
example_repo_setup(files_to_remove = basename(example_files))
}
