% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gs_png.R
\name{gs_png_url}
\alias{gs_png_url}
\title{Title}
\usage{
gs_png_url(id, slide_id)
}
\arguments{
\item{slide_id}{}
}
\value{

}
\description{
Title
}
\examples{
url = paste0("https://docs.google.com/presentation/d/",
"12DPZgPteQBwgal6kSPP58zhPhjZ7QSPZLe3NkA8M3eo/edit",
"#slide=id.gc8648f14c3_0_397&t=4")
id = ariExtra::get_slide_id(url)
slide_id = get_slide_page(url)
gs_png_url(id, slide_id)
}
