% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bookdown_to_leanpub.R
\name{make_embed_markdown}
\alias{make_embed_markdown}
\title{Make Leanpub file that has embed webpage of a chapter}
\usage{
make_embed_markdown(
  url,
  chapt_title,
  img_path,
  output_dir = "manuscript",
  verbose = TRUE
)
}
\arguments{
\item{url}{The url to the chapter that is to be embed}

\item{chapt_name}{Title of chapter to be used as file name and printed on iframe}

\item{img}{File path to image to use for poster}
}
\value{
A
}
\description{
Make Leanpub file that has embed webpage of a chapter
}
