% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quiz.R
\name{parse_quiz}
\alias{parse_quiz}
\alias{extract_quiz}
\alias{find_quiz_indices}
\title{Parse Quiz}
\usage{
parse_quiz(x)

extract_quiz(x)

find_quiz_indices(x)
}
\arguments{
\item{x}{A single filename or a vector of the contents of the markdown
file}
}
\value{
A list of elements, including a `data.frame` and metadata
for questions
}
\description{
Parse Quiz
}
\examples{

x = c('{quiz, id: quiz_00_filename}',
"### Lesson Name quiz",
"{choose-answers: 4}",
"? What do you think?",
"C) The answer to this one",
"o) Not the answer",
"o) Not the answer either",
"C) Another correct answer",
"m) Mandatory different answer",
"{/quiz}")
out = parse_quiz(x)
check_quiz_attributes(out)
}
