% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/replace_html.R
\name{replace_html}
\alias{replace_html}
\alias{replace_single_html}
\title{Replace HTML and other Tags in Leanpub Markdown}
\usage{
replace_html(
  path = "manuscript",
  remove_resources_start = TRUE,
  footer_text = NULL,
  fullbleed = FALSE,
  verbose = TRUE
)

replace_single_html(
  file,
  footer_text = NULL,
  remove_resources_start = TRUE,
  fullbleed = FALSE,
  verbose = TRUE
)
}
\arguments{
\item{path}{path to the markdown files that need replacement.}

\item{remove_resources_start}{remove the word `resources/` at the front
of any image path.}

\item{footer_text}{a bit of text that will be added to the
end of each file before the references section.}

\item{fullbleed}{should the image have the attribute `fullbleed: true`?}

\item{verbose}{print diagnostic messages}

\item{file}{individual markdown file}
}
\value{
A list of output files and diagnostics
}
\description{
Replace HTML and other Tags in Leanpub Markdown
}
