% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/replace_html.R
\name{replace_html}
\alias{replace_html}
\title{Replace HTML and other Tags in Leanpub Markdown}
\usage{
replace_html(path = "manuscript", verbose = TRUE)
}
\arguments{
\item{path}{path to the bookdown book, must have a `_bookdown.yml` file}

\item{verbose}{print diagnostic messages}

\item{output_dir}{output directory to put files.  It should likely be
relative to path}
}
\value{
A list of output files and diagnostics
}
\description{
Replace HTML and other Tags in Leanpub Markdown
}
