% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{set_knitr_image_path}
\alias{set_knitr_image_path}
\title{set_knitr_image_path is being deprecated}
\usage{
set_knitr_image_path()
}
\value{
Message that says to no longer use this function
}
\description{
set_knitr_image_path is being deprecated
}
