% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_knitr_image_path.R
\name{set_knitr_image_path}
\alias{set_knitr_image_path}
\title{Set image path for `knitr`}
\usage{
set_knitr_image_path(verbose = FALSE)
}
\arguments{
\item{verbose}{print out what the figure path is}
}
\description{
Set image path for `knitr`
}
